bugfix large zoom

+++ src/mmhp.cpp
@@ -30,8 +30,8 @@ int main(int argc, char* argv[])
 	static FltH    zoom_step       = from_stringH ( cmd.get_parameter ( "zoom-step",    "0.99"    ));
 	static FltH    center_x        = from_stringH ( cmd.get_parameter ( "center-x",     "0.5"     ));
 	static FltH    center_y        = from_stringH ( cmd.get_parameter ( "center-y",     "0.0"     ));
-	static float   mod_base        = std::stof    ( cmd.get_parameter ( "col-base",     "100"     ));
-	static float   mod_pow         = std::stof    ( cmd.get_parameter ( "col-pow",      "0.03"    ));
+	static double  mod_base        = std::stod    ( cmd.get_parameter ( "col-base",     "100"     ));
+	static double  mod_pow         = std::stod    ( cmd.get_parameter ( "col-pow",      "0.03"    ));
 	static UL      num_dig         = std::stol    ( cmd.get_parameter ( "num-digits",   "3"       ));
 	static UL      update_cap      = std::stol    ( cmd.get_parameter ( "update-cap",   "50"      ));
 	static UL      image_width     = std::stol    ( cmd.get_parameter ( "width",        "640"     ));
@@ -185,7 +185,7 @@ int main(int argc, char* argv[])
 		logger << "scale-x        : " << mh.scale_x << std::endl;
 		logger << "scale-y        : " << mh.scale_y << std::endl;
 
-		ModFunc mod_func = [](float d) -> float { return mod_base / powf(d, mod_pow); };
+		ModFunc mod_func = [](double d) -> float { return mod_base / std::pow(d, mod_pow); };
 		UL maxout;
 
 		auto EXEC = [&](int n) -> void
@@ -197,7 +197,7 @@ int main(int argc, char* argv[])
 			auto d1 = std::chrono::duration_cast<std::chrono::milliseconds>(t2-t1).count();
 			logger << "generate time  : " << d1 / 1000.0f << std::endl;
 			logger << "effective cap  : " << maxout << std::endl;
-			logger << "color mod      : " << mod_func((float)(double)zoom_cur) << std::endl;
+			logger << "color mod      : " << mod_func((double)zoom_cur) << std::endl;
 			for (int j=0; j<n; ++j)
 			{
 				curr_name = mkname(i+j);
@@ -263,7 +263,7 @@ int main(int argc, char* argv[])
 			auto t2 = std::chrono::high_resolution_clock::now();
 			auto d1 = std::chrono::duration_cast<std::chrono::milliseconds>(t2-t1).count();
 			logger << "generate time  : " << d1 / 1000.0f << std::endl;
-			logger << "color mod      : " << mod_func((float)(double)zoom_cur) << std::endl;
+			logger << "color mod      : " << mod_func((double)zoom_cur) << std::endl;
 			auto zc = zoom_cur;
 
 			/* */
